+++ web/components/SearchBar.vue
